[replace] app module
